Removed custom migrations and MigrationInstaller

+++ bundles/CoreBundle/Migrations/Version20201012154224.php
@@ -3,22 +3,24 @@
 namespace Pimcore\Bundle\CoreBundle\Migrations;
 
 use Doctrine\DBAL\Schema\Schema;
-use Pimcore\Migrations\Migration\AbstractPimcoreMigration;
+use Doctrine\Migrations\AbstractMigration;
 
-class Version20201012154224 extends AbstractPimcoreMigration
+class Version20201012154224 extends AbstractMigration
 {
     /**
      * @param Schema $schema
      */
-    public function up(Schema $schema)
+    public function up(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE glossary DROP COLUMN acronym');
+        if($schema->getTable('glossary')->hasColumn('acronym')) {
+            $this->addSql('ALTER TABLE glossary DROP COLUMN acronym');
+        }
     }
 
     /**
      * @param Schema $schema
      */
-    public function down(Schema $schema)
+    public function down(Schema $schema): void
     {
         $this->addSql('ALTER TABLE glossary ADD COLUMN `acronym` varchar(255) DEFAULT NULL');
     }
